Created a Free-Assign mapping for LC XL on Bitwig
This first version assigns the eight user modes of the LaunchControlXL
to any selected control in Bitwig. Just create a controller assignment
in Bitwig, tweak a control and... you're all set.

This version supports eight freely-assignable user boards, more than
enough for a long LiveSet. Factory modes, more appropriate for
composing, editing and jamming in general will come in the next release.

Since it's possible to switch between control boards, a soft-takeover
algorithm has been implemented on the faders and knobs. Feedback is
provided to the user thanks to the leds: a switched off led means that
the control is currently unassigned, a full led means that you're in
control and a blinking led means that the knob is in soft-takeover mode
and has to be tweaked until it reaches the value it has in the software
to be active again.

The code itself has been designed in a way that will make it easier to
implement Factory modes. Every new (factory) board will also inherit
from SoftTakeoverBoard (since it has to implement that feature too) and
will have to forward midi events, software events and enable/disable
events to the latter in a way that is described in the code.
The getSoftValue() and setSoftValue() methods will obviously have to be
overriden as well.
Ultimately, navigation can be made more comprehensive if Factory boards
overide the get(Weak)ColorBits() function to actually show some kind of
"scrolling" feedback.