Add abstraction for a mixer board
Device mode and knobs don't do anything on such a board. That's why I
called it an "abstraction" but technically speaking it can still be
instianciated.

It comes with track scrolling features, solo/mute/arm and volume
controls. The select feature is yet to come.

Will follow an abstraction for a DeviceBoard with device mode (but still
no faders or vertical scrolling in Mixer mode) and finally, several
child boards such as the macro board, the live (free assign knobs)
boards (there must be at least two : one for drums and one for other
stuff, maybe one more just in case) or the FX (+pan) board.