Add select feature and Track color following
The later projects the color of each channel on the poor launchpad color
set and assigns this color to the select button. Only bad thing, the
default color is kind of awful and it makes the highlighting of the
selected track less obvious. Maybe we should improve the projection of
the pale shade by having a prefined one for each light shade. This way
we could have a more uniform repartition of colors.

Or even better, only project among a set of 2x 4 shades we're using for
the user mode of each channel on the poor launchpad color set and
assigns this color to the select button. Only bad thing, the default
color is kind of awful and it makes the highlighting of the selected
track less obvious. Maybe we should improve the projection of the pale
shade by having a prefined one for each light shade. This way we could
have a more uniform repartition of colors.

Or even better, only project among a set of 2x 4 shades we're using for
the user mode.

Some bugs have also been fixed and we now detect unexisiting tracks :)