Add Macro board : a first complete factory mode
That's it ! The macro board is here. It provides control over the macros
of the primary device of each channel with "vertical" scrolling to have
access to the eight macro. Complete soft takeover support is here,
whatever move you want to achieve on the board. The implementation for
the pan/fx board and live board are on their way now!