Here comes the FX sends/pan board
It's just another classic mixer board with the bottom row controlling
the instruments pan and the two on top controlling FX sends.
Unfortunately, vertical scrolling isn't working in my version of Bitwig
even though it should be implemented on my end. It might be a bug in
Bitwig's controller API. I'm stuck on 1.0.15 for now but I'll upgrade
and try with a more recent version. If it doesn't work out of the box
I'll find a workaround but I don't want to waste time implementing it
now if it works out of the box on newer versions.

Has also been added a feature to have visual feedback about which
controls are actually mapped to the current board layout.

Very soon will come the live board, which will be the occasion of a
last 0.x release before the device board is there and some testing is
done (this will be 1.0.0).

Plans for the future: also having a mixer board where buttons buttons
behaviour is rewritten to have control over transport, Bitwig's basic
features (like undo/redo) and navigation through tracks/presets all of
it with the possibility to insert and select a new track with the
instrument being highlighted in that view.