And here comes the device modes for mixer boards!
It has a couple of nice features, everything is taking over accordingly.
The top row of knobs is mapped to a device's common parameters, the
bottom one to it's macros and the middle one is mapped to the
instrument's pages of control, which can be scrolled through with the
up/down navigation buttons.

The faders control the envelope parameters.

Scrolling in a device chain is made possible by the left/right
navigation buttons.

The action buttons are unused for now but are reserved for in-depth (or
z-axis, or nested) device navigation (up/down and selected nested device
for all devices except containers where we could use the pad buttons to
select a subdevice. I'm thinking about the Drum Machine in particular).

Finally, the first four columns of buttons are dedicated to
enabling/disabling macro modulation on the device. In that mode the
other controls can be used to set the modulation amount, the fifth
column provides buttons to arm/unarm the track the device belongs to and
to toggle the device's enabled state and the last buttons enable
navigation accross preset category, author and preset themselves. Use we
care if you don't want to lose a device config by accident :)