Soft-takeover mode is now using weak colors
Instead of blinking which was producinga lot of noise when many leds
were blinking at the same time. I'm not completely sure this will stay
because it's less obvious which leds are actually taking over and which
aren't. I might as well try to reduce the brightness off the weak colors
but since we have control over 4 bits only to control the leds (color
AND brightness), it might be tough.