Fix soft-takeover led feedback bug
Due to devices calling `SoftTakeoverBoard.valueChangedCallback` in mixer
modes and vice versa, there was a bug with soft-takeover and led
feedback. If you changed the third macro on the selected device on the
macro board, the third led on the `Pan/Device`  would have had a random
behaviour. This is know fixed.